Fix path and fix double checking of circular contigs
First, fix path for checking for circular contigs.
Unicycler.sh is expecting the cir rep contigs file
in the output dir, not the unicycler dir.

Also fix split.sh removing the cir_rep_contigs.fasta file in
the output dir instead of the unicycler dir. If the file is left
in the unicycler dir, then the *.fasta glob in unicycler.sh will
grab not only the split contig files, but also the cir_rep_contig
file, double counting them.